Moved ApplicationReviewAndSubmitComponent
