+++ app/views/candidate_interface/continuous_applications/course_choices/review/show.html.erb
@@ -9,6 +9,6 @@
 
     <%= render(CandidateInterface::ContinuousApplications::ApplicationReviewComponent.new(application_choice: @application_choice)) %>
 
-    <%= render(CandidateInterface::ContinuousApplications::ApplicationReviewAndSubmitComponent.new(application_choice: @application_choice)) %>
+    <%= render(CandidateInterface::ApplicationReviewAndSubmitComponent.new(application_choice: @application_choice)) %>
   </div>
 </div>
